+++ backend/scripts/map-gen/command-disc.ts
@@ -1,7 +1,7 @@
 import { createRng, mulberry32Prng } from "@guillaume-docquier/tools-ts"
 import { Command } from "commander"
 import { discGenerator } from "#lib/map/points/disc.generator.ts"
-import type { Point2D } from "#lib/map/points/Point.ts"
+import type { Point2D } from "#lib/map/points/Point2D.ts"
 import { randomUInt32 } from "#lib/randomUInt32.ts"
 import { Parser } from "./Parser.ts"
 import { SvgRenderer } from "./SvgRenderer.ts"
