+++ backend/src/lib/map/system.generator.ts
@@ -1,5 +1,5 @@
 import type { Planet } from "#lib/map/planet.generator.ts"
-import type { Point2D } from "#lib/map/points/Point.ts"
+import type { Point2D } from "#lib/map/points/Point2D.ts"
 import { type Star, starGenerator } from "#lib/map/star.generator.ts"
 
 export type System = {
